stab/requirements.txt
fastapi
uvicorn[standard]
requests
mgrs
python-dotenv
